split gui class in events, listeners and controller
