+++ src/javascript/index.js
@@ -1,41 +1,7 @@
 /**
  * executes all needed classes and functions for running the logic of path-finder
  */
-let Gui = require("./Gui");
+let GuiController = require("./GuiController");
 
-let gui = new Gui();
-
-let grid = createGrid(24, 24, function (el, row, col, i) {
-  el.className = "clicked";
-  if (gui.choosingObstacle) {
-    gui.setObstacle(col, row);
-  } else if (gui.choosingEndPoint) {
-    gui.setEndPoint(col, row);
-  } else {
-    gui.setStartPoint(col, row);
-  }
-});
-
-function createGrid(rows, cols, callback) {
-  var i = 0;
-  var grid = document.createElement("table");
-  grid.className = "grid";
-  for (var r = 0; r < rows; ++r) {
-    var tr = grid.appendChild(document.createElement("tr"));
-    for (var c = 0; c < cols; ++c) {
-      var cell = tr.appendChild(document.createElement("td"));
-      cell.addEventListener(
-        "click",
-        (function (el, r, c, i) {
-          return function () {
-            callback(el, r, c, i);
-          };
-        })(cell, r, c, i),
-        false
-      );
-    }
-  }
-  return grid;
-}
-
-gui.setGrid(grid);
+let guiController = new GuiController();
+guiController.setup();
